Move API analysis utilities to own file
